+++ packages/mpack/src/experimental/server/DevServer.ts
@@ -278,6 +278,5 @@ export class DevServer {
 }
 
 declare global {
-  // eslint-disable-next-line no-var
   var remoteBundles: Record<'android' | 'ios', string> | null;
 }