+++ packages/mpack/CHANGELOG.md
@@ -1,5 +1,85 @@
 # @granite-js/mpack
 
+## 0.1.34-next.8
+
+### Patch Changes
+
+- fix
+- Updated dependencies
+  - @granite-js/devtools-frontend@0.1.34-next.8
+  - @granite-js/plugin-core@0.1.34-next.8
+  - @granite-js/utils@0.1.34-next.8
+
+## 0.1.34-next.7
+
+### Patch Changes
+
+- fix
+- Updated dependencies
+  - @granite-js/devtools-frontend@0.1.34-next.7
+  - @granite-js/plugin-core@0.1.34-next.7
+  - @granite-js/utils@0.1.34-next.7
+
+## 0.1.34-next.6
+
+### Patch Changes
+
+- fix
+- Updated dependencies
+  - @granite-js/devtools-frontend@0.1.34-next.6
+  - @granite-js/plugin-core@0.1.34-next.6
+  - @granite-js/utils@0.1.34-next.6
+
+## 0.1.34-next.5
+
+### Patch Changes
+
+- fix
+- Updated dependencies
+  - @granite-js/devtools-frontend@0.1.34-next.5
+  - @granite-js/plugin-core@0.1.34-next.5
+  - @granite-js/utils@0.1.34-next.5
+
+## 0.1.34-next.4
+
+### Patch Changes
+
+- fix
+- Updated dependencies
+  - @granite-js/devtools-frontend@0.1.34-next.4
+  - @granite-js/plugin-core@0.1.34-next.4
+  - @granite-js/utils@0.1.34-next.4
+
+## 0.1.34-next.3
+
+### Patch Changes
+
+- fix
+- Updated dependencies
+  - @granite-js/devtools-frontend@0.1.34-next.3
+  - @granite-js/plugin-core@0.1.34-next.3
+  - @granite-js/utils@0.1.34-next.3
+
+## 0.1.34-next.2
+
+### Patch Changes
+
+- fix
+- Updated dependencies
+  - @granite-js/devtools-frontend@0.1.34-next.2
+  - @granite-js/plugin-core@0.1.34-next.2
+  - @granite-js/utils@0.1.34-next.2
+
+## 0.1.34-next.1
+
+### Patch Changes
+
+- fix
+- Updated dependencies
+  - @granite-js/devtools-frontend@0.1.34-next.1
+  - @granite-js/plugin-core@0.1.34-next.1
+  - @granite-js/utils@0.1.34-next.1
+
 ## 0.1.34-next.0
 
 ### Patch Changes