+++ packages/plugin-core/CHANGELOG.md
@@ -1,5 +1,69 @@
 # @granite-js/plugin-core
 
+## 0.1.34-next.8
+
+### Patch Changes
+
+- fix
+- Updated dependencies
+  - @granite-js/utils@0.1.34-next.8
+
+## 0.1.34-next.7
+
+### Patch Changes
+
+- fix
+- Updated dependencies
+  - @granite-js/utils@0.1.34-next.7
+
+## 0.1.34-next.6
+
+### Patch Changes
+
+- fix
+- Updated dependencies
+  - @granite-js/utils@0.1.34-next.6
+
+## 0.1.34-next.5
+
+### Patch Changes
+
+- fix
+- Updated dependencies
+  - @granite-js/utils@0.1.34-next.5
+
+## 0.1.34-next.4
+
+### Patch Changes
+
+- fix
+- Updated dependencies
+  - @granite-js/utils@0.1.34-next.4
+
+## 0.1.34-next.3
+
+### Patch Changes
+
+- fix
+- Updated dependencies
+  - @granite-js/utils@0.1.34-next.3
+
+## 0.1.34-next.2
+
+### Patch Changes
+
+- fix
+- Updated dependencies
+  - @granite-js/utils@0.1.34-next.2
+
+## 0.1.34-next.1
+
+### Patch Changes
+
+- fix
+- Updated dependencies
+  - @granite-js/utils@0.1.34-next.1
+
 ## 0.1.34-next.0
 
 ### Patch Changes